14.2.5 Listing 14.34-7: Use AJAX for `follow/unfollow` buttons
`remote: true` in a form button allows the request to be sent to the server
without forcing a page refresh.

A variable is added to the form tag, `data-remote="true"`, which allows the request
to be handled by JavaScript, instead of as a standard request within the browser.

The `create` and `destroy` actions use the `respond_to` block to control how the
current page is updated.

  - If the page is only HTML, then redirect to the user page
  - If the page has JS enabled, then update without a full refresh

Finally, update `application.rb` to put `authenticity_token` in `remote` (JS controlled) forms.

This won't work yet, because the JS files don't exist yet.